pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insti-scraper"
version = "0.1.0"
description = "AI-powered universal faculty data scraper for universities"
readme = "insti_scraper/README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "InstiGPT Team" }
]
keywords = ["scraper", "faculty", "university", "crawl4ai", "llm"]

dependencies = [
    "crawl4ai>=0.7.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "cloudscraper>=1.2.0",
    "litellm>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "ddgs>=9.10.0",
    "pillow>=12.1.0",
    "sqlmodel>=0.0.31",
    "rich>=14.2.0",
    "alembic>=1.18.1",
    "httpx>=0.28.1",
    "google-generativeai>=0.3.0; extra == 'vision'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.4.0",
]
vision = [
    "google-generativeai>=0.3.0",
]

[project.scripts]
insti-scraper = "insti_scraper.main:main"
insti-batch = "insti_scraper.batch:main"

[tool.hatch.build.targets.wheel]
packages = ["insti_scraper"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
